% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashsvd.R
\name{dashSVD}
\alias{dashSVD}
\title{Dynamic shifted Randomized SVD}
\usage{
dashSVD(A, k, p = 10, s = 10)
}
\description{
Dynamic shifted Randomized SVD
}
\examples{
mat <- matrix(rnorm(100*20000), 20000, 100)
res <- dashSVD(mat, k = 10)
str(res)
}
