% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCAoneAlg1}
\alias{PCAoneAlg1}
\title{The algorithm1 in PCAone paper.}
\usage{
PCAoneAlg1(mat, k, p, q, rand, finder = 1L)
}
\arguments{
\item{mat}{the input matrix.}

\item{k}{top k singular values, k << any(dim(mat)).}

\item{p}{number of power iterations.}

\item{q}{oversampling.}

\item{rand}{distribution of random matrix. 1: standard noraml distribution. 2: uniform distribution}

\item{finder}{method to find othogonal matrix. 1: QR. 2: LU}
}
\description{
The algorithm1 in PCAone paper.
}
