% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCAoneAlg2}
\alias{PCAoneAlg2}
\title{The algorithm1 in PCAone paper.}
\usage{
PCAoneAlg2(mat, k, p, q, rand, windows)
}
\arguments{
\item{mat}{the input matrix.}

\item{k}{top k singular values, k << any(dim(mat)).}

\item{p}{number of power iterations.}

\item{q}{oversampling.}

\item{rand}{distribution of random matrix. 1: standard noraml distribution. 2: uniform distribution}

\item{windows}{the number of windows. must be a power of 2}
}
\description{
The algorithm1 in PCAone paper.
}
