% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCAoneH}
\alias{PCAoneH}
\title{one pass randomized svd in PCAone, adopted from Yu's paer.}
\usage{
PCAoneH(mat, k, l, p)
}
\arguments{
\item{mat}{the input matrix.}

\item{k}{top k singular values, k << any(dim(mat)).}

\item{l}{oversampling.}

\item{p}{number of power iterations.}
}
\description{
one pass randomized svd in PCAone, adopted from Yu's paer.
}
