% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCAoneLi}
\alias{PCAoneLi}
\title{PCAone: the window-based one pass randomized svd, Li et al 2022.}
\usage{
PCAoneLi(mat, k, p, q, rand, windows)
}
\arguments{
\item{mat}{the input matrix.}

\item{k}{top k singular values, k << any(dim(mat)).}

\item{p}{number of power iterations.}

\item{q}{oversampling.}

\item{rand}{distribution of random matrix. 1: standard noraml distribution. 2: uniform distribution}

\item{windows}{the number of windows. must be a power of 2}
}
\description{
PCAone: the window-based one pass randomized svd, Li et al 2022.
}
