% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCAoneYu}
\alias{PCAoneYu}
\title{The one pass randomized svd in PCAone, modified from Yu et al 2017.}
\usage{
PCAoneYu(mat, k, p, q, rand, finder)
}
\arguments{
\item{mat}{the input matrix.}

\item{k}{top k singular values, k << any(dim(mat)).}

\item{p}{number of power iterations.}

\item{q}{oversampling.}

\item{rand}{distribution of random matrix. 1: standard noraml distribution. 2: uniform distribution}
}
\description{
The one pass randomized svd in PCAone, modified from Yu et al 2017.
}
