\docType{data}
\name{popgen}
\alias{popgen}
\title{Population genetic data}
\format{
An object of class \code{\link[pcaone]{pcaone}}.
}
\source{
\url{https://www.internationalgenome.org}
}
\usage{
load(system.file("extdata", "popgen.rda", package="pcaone") )
}
\description{
Four populations from East Asia in the 1000 Genome Project
}
\keyword{datasets}
