% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsvd.R
\name{winSVD}
\alias{winSVD}
\title{Window-based Randomized SVD}
\usage{
winSVD(A, k, p = 10, s = 10, B = 64, perm = 1)
}
\arguments{
\item{A}{array_like; \cr
a real \eqn{(m, n)} input matrix to be decomposed.}

\item{k}{integer; \cr
specifies the target rank of the low-rank decomposition. \eqn{k} should satisfy \eqn{k << min(m,n)}.}

\item{p}{integer, optional; \cr
number of additional power iterations (by default \eqn{p=10}).}

\item{s}{integer, optional; \cr
oversampling parameter (by default \eqn{s=10}).}

\item{B}{integer, optional; \cr
number of windows (by default \eqn{B=64}).}

\item{perm}{integer, optional; \cr
the strategy for pivoting/permutation of rows : \cr
\eqn{'1'} :  random (default)  \cr
\eqn{'2'} :  partial random  \cr
\eqn{'0'} :  no permutation  \cr}
}
\value{
\code{winSVD} returns a list containing the following three components:
\describe{
\item{d}{  array_like; \cr
singular values; vector of length \eqn{(k)}.
}

\item{u}{  array_like; \cr
left singular vectors; \eqn{(m, k)} or \eqn{(m, nu)} dimensional array.
}

\item{v}{  array_like; \cr
right singular vectors; \eqn{(n, k)} or \eqn{(n, nv)} dimensional array. \cr
}
}
}
\description{
winSVD implements the window-based Randomized SVD proposed by Li et al. 2023
}
\note{
The singular vectors are not unique and only defined up to sign.
If a left singular vector has its sign changed, changing the sign of the corresponding right vector
gives an equivalent decomposition.
}
\examples{
mat <- matrix(rnorm(100*20000), 20000, 100)
res <- winSVD(mat, k = 10)
str(res)
}
\references{
\itemize{
\item Z. Li, J Meisner, A Albrechtsen. "Fast and accurate out-of-core PCA framework for large scale biobank data" (2023)
\doi{10.1101/gr.277525.122}.
}
}
\author{
Zilong Li \email{zilong.dk@gmail.com}
}
