% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popgen}
\alias{popgen}
\title{Population genetic data}
\format{
An object of class \code{\link[pcaone]{pcaone}}.
}
\source{
\href{https://www.internationalgenome.org/}
}
\usage{
data('popgen')
}
\description{
Four populations from East Asia in the 1000 Genome Project
}
\examples{
\dontrun{
library('pcaone')
data('popgen')
}
}
\keyword{datasets}
